Refactor cyral_role and cyral_rego_policy_instance
