+++ cyral/internal/regopolicy/constants.go
@@ -1,0 +1,26 @@
+package regopolicy
+
+const (
+	resourceName = "cyral_rego_policy_instance"
+)
+
+const (
+	// Schema keys
+	RegoPolicyInstanceResourceIDKey  = "id"
+	RegoPolicyInstancePolicyIDKey    = "policy_id"
+	RegoPolicyInstanceCategoryKey    = "category"
+	RegoPolicyInstanceNameKey        = "name"
+	RegoPolicyInstanceDescriptionKey = "description"
+	RegoPolicyInstanceTemplateIDKey  = "template_id"
+	RegoPolicyInstanceParametersKey  = "parameters"
+	RegoPolicyInstanceEnabledKey     = "enabled"
+	RegoPolicyInstanceScopeKey       = "scope"
+	RegoPolicyInstanceRepoIDsKey     = "repo_ids"
+	RegoPolicyInstanceTagsKey        = "tags"
+	RegoPolicyInstanceDurationKey    = "duration"
+	RegoPolicyInstanceLastUpdatedKey = "last_updated"
+	RegoPolicyInstanceCreatedKey     = "created"
+	RegoPolicyInstanceActorKey       = "actor"
+	RegoPolicyInstanceActorTypeKey   = "actor_type"
+	RegoPolicyInstanceTimestampKey   = "timestamp"
+)
