+++ cyral/internal/role/constants.go
@@ -1,0 +1,23 @@
+package role
+
+const (
+	resourceName   = "cyral_role"
+	dataSourceName = "cyral_role"
+)
+
+const (
+	// Schema keys
+	approvalManagementPermissionKey           = "approval_management"
+	modifyPoliciesPermissionKey               = "modify_policies"
+	modifyRolesPermissionKey                  = "modify_roles"
+	modifySidecarAndRepositoriesPermissionKey = "modify_sidecars_and_repositories"
+	modifyUsersPermissionKey                  = "modify_users"
+	repoCrawlerPermissionKey                  = "repo_crawler"
+	viewAuditLogsPermissionKey                = "view_audit_logs"
+	viewDatamapsPermissionKey                 = "view_datamaps"
+	viewIntegrationsPermissionKey             = "view_integrations"
+	viewPoliciesPermissionKey                 = "view_policies"
+	viewRolesPermissionKey                    = "view_roles"
+	viewUsersPermissionKey                    = "view_users"
+	modifyIntegrationsPermissionKey           = "modify_integrations"
+)
