+++ cyral/internal/role/resource.go
@@ -15,24 +15,14 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-// Roles correspond to Groups in API.
-type RoleDataRequest struct {
-	Name string `json:"name,omitempty"`
-	// Permissions correspond to Roles in API.
-	PermissionIDs []string `json:"roles,omitempty"`
+var urlFactory = func(d *schema.ResourceData, c *client.Client) string {
+	return fmt.Sprintf("https://%s/v1/users/groups/%s", c.ControlPlane, d.Id())
 }
 
-// Roles correspond to Groups in API.
-type RoleDataResponse struct {
-	Id   string `json:"id,omitempty"`
-	Name string `json:"name,omitempty"`
-	// Permissions correspond to Roles in API.
-	Permissions []*permission.Permission `json:"roles,omitempty"`
-}
-
-func ResourceRole() *schema.Resource {
+func resourceSchema() *schema.Resource {
 	return &schema.Resource{
-		Description:   "Manages [roles for Cyral control plane users](https://cyral.com/docs/account-administration/acct-manage-cyral-roles/#create-and-manage-administrator-roles-for-cyral-control-plane-users). See also: [Role SSO Groups](./role_sso_groups.md).",
+		Description: "Manages [roles for Cyral control plane users](https://cyral.com/docs/account-administration/acct-manage-cyral-roles/#create-and-manage-administrator-roles-for-cyral-control-plane-users). See also: [Role SSO Groups](./role_sso_groups.md).",
+
 		CreateContext: resourceRoleCreate,
 		ReadContext:   resourceRoleRead,
 		UpdateContext: resourceRoleUpdate,
@@ -55,7 +45,88 @@ func ResourceRole() *schema.Resource {
 				Optional:    true,
 				MaxItems:    1,
 				Elem: &schema.Resource{
-					Schema: permissionsSchema,
+					Schema: map[string]*schema.Schema{
+						approvalManagementPermissionKey: {
+							Description: "Allows approving or denying approval requests on Cyral Control Plane. " +
+								"Defaults to `false`.",
+							Type:     schema.TypeBool,
+							Optional: true,
+							Default:  false,
+						},
+						modifyPoliciesPermissionKey: {
+							Description: "Allows modifying policies on Cyral Control Plane. Defaults to `false`.",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+						},
+						modifyRolesPermissionKey: {
+							Description: "Allows modifying roles on Cyral Control Plane. Defaults to `false`.",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+						},
+						modifySidecarAndRepositoriesPermissionKey: {
+							Description: "Allows modifying sidecars and repositories on Cyral Control Plane. Defaults to `false`.",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+						},
+						modifyUsersPermissionKey: {
+							Description: "Allows modifying users on Cyral Control Plane. Defaults to `false`.",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+						},
+						repoCrawlerPermissionKey: {
+							Description: "Allows running the Cyral repo crawler data classifier and user discovery. " +
+								"Defaults to `false`.",
+							Type:     schema.TypeBool,
+							Optional: true,
+							Default:  false,
+						},
+						viewAuditLogsPermissionKey: {
+							Description: "Allows viewing audit logs on Cyral Control Plane. Defaults to `false`.",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+						},
+						viewDatamapsPermissionKey: {
+							Description: "Allows viewing datamaps on Cyral Control Plane. Defaults to `false`.",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+						},
+						viewIntegrationsPermissionKey: {
+							Description: "Allows viewing integrations on Cyral Control Plane. Defaults to `false`.",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+						},
+						viewPoliciesPermissionKey: {
+							Description: "Allows viewing policies on Cyral Control Plane. Defaults to `false`.",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+						},
+						viewRolesPermissionKey: {
+							Description: "Allows viewing roles on Cyral Control Plane. Defaults to `false`.",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+						},
+						viewUsersPermissionKey: {
+							Description: "Allows viewing users on Cyral Control Plane. Defaults to `false`.",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+						},
+						modifyIntegrationsPermissionKey: {
+							Description: "Allows modifying integrations on Cyral Control Plane. Defaults to `false`.",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+						},
+					},
 				},
 			},
 		},
