+++ cyral/internal/role/ssogroups/constants.go
@@ -1,0 +1,5 @@
+package ssogroups
+
+const (
+	resourceName = "cyral_role_sso_groups"
+)
